refactor and minor fixes
