+++ openproblems/tasks/spatial_decomposition/_utils.py
