+++ openproblems/tasks/spatial_decomposition/methods/nnls.py
@@ -1,7 +1,7 @@
 from ....tools.decorators import method
 from ....tools.utils import check_version
-from ._utils import normalize_coefficients
-from ._utils import obs_means
+from .._utils import normalize_coefficients
+from .._utils import obs_means
 from scipy.optimize import nnls
 from scipy.sparse import issparse
