+++ openproblems/tasks/spatial_decomposition/methods/vanillanmf.py
@@ -1,24 +1,10 @@
 from ....tools.decorators import method
 from ....tools.utils import check_version
-from anndata import AnnData
+from .._utils import obs_means
 from scipy.sparse import issparse
 from sklearn.decomposition import NMF
 
 import numpy as np
-
-
-def obs_means(adata: AnnData, cluster_key: str) -> AnnData:
-    """Return means over observation key."""
-
-    labels = adata.obs[cluster_key].cat.categories
-    means = np.empty((labels.shape[0], adata.shape[1]))
-    for i, lab in enumerate(labels):
-        means[i, :] = adata[adata.obs[cluster_key] == lab].X.mean(axis=0).flatten()
-    adata_means = AnnData(means)
-    adata_means.obs_names = labels
-    adata_means.var_names = adata.var_names
-
-    return adata_means
 
 
 @method(
